Refactor core base types
ExperimentSession is refactored to inherit from aeon base and removed
redundant definitions.
