+++ src/ucl_open/_generators/json_schema.py
@@ -3,7 +3,21 @@ from typing import Union
 import json
 import pydantic
 
-from ucl_open.core.schema import DataTypes
+import ucl_open.core.data_types as _core_dt
+import ucl_open.devices.data_types as _dev_dt
+
+
+def _collect_fields(*modules):
+    fields = {}
+    for mod in modules:
+        for name in getattr(mod, "__all__", []):
+            cls = getattr(mod, name)
+            if isinstance(cls, type):
+                fields[name] = (cls, ...)
+    return fields
+
+
+DataTypes = pydantic.create_model("DataTypes", **_collect_fields(_core_dt, _dev_dt))
 
 SCHEMA_ROOT = Path("./src/ucl_open/schemas")
