+++ src/ucl_open/core/experiment.py
@@ -1,12 +1,9 @@
 from pydantic import Field
-from ucl_open.core.base import BaseSchema
+from swc.aeon.schema import Experiment
 
 
-class Experiment(BaseSchema):
+class ExperimentSession(Experiment):
     """The base class for creating ucl-open experiment models."""
 
-    workflow: str = Field(description="Path to the workflow running the experiment.")
-    commit: str = Field(description="Commit hash of the experiment/rig repo.")
-    repository_url: str = Field(
-        description="The URL of the git repository used to version experiment source code."
-    )
+    subject_id: str = Field(description="The subject id for this session.")
+    session_id: str = Field(description="The session identifier.")
