+++ src/ucl_open/devices/data_types.py
@@ -1,6 +1,6 @@
 from typing import Dict
 from pydantic import Field
-from ucl_open.core.base import BaseSchema
+from ucl_open.core import BaseSchema
 
 
 class StepperPositions(BaseSchema):
